FS-B-7

+++ backend/FileService/src/FileService.Core/MediaAssets/IChunkSizeCalculator.cs
@@ -5,10 +5,10 @@ namespace FileService.Core.MediaAssets;
 
 public interface IChunkSizeCalculator
 {
-    Result<(long ChuckSize, int TotalChunks), Error> Calculate(
+    Result<(int ChuckSize, int TotalChunks), Error> Calculate(
         long fileSize,
         long recommendedChunksSizeBytes,
         int maxChunks);
 
-    Result<(long ChuckSize, int TotalChunks), Error> Calculate(long fileSize);
+    Result<(int ChuckSize, int TotalChunks), Error> Calculate(long fileSize);
 }
